Clean up node code. Trying to get things to work.
