--- profile.html
@@ -1,13 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>Title</title>
-</head>
-<body>
-<form id="profileform" action="/users/1" method="POST">
-    Name: <input id="profilename" name="name" type="text" />
-</form>
-
-</body>
-</html>
