+++ public/profile.html
@@ -1,0 +1,14 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <script src="profile.js"> </script>
+    <meta charset="UTF-8">
+    <title>Title</title>
+</head>
+<body>
+<form id="profileform" action="/users/1" method="POST">
+    Name: <input id="profilename" name="name" type="text" />
+</form>
+
+</body>
+</html>
